# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PGP Key Manager API
  version: 1.0.0
  contact:
    name: PGP Key Manager
    url: https://github.com/bbruneel/PGP-key-manager
  license:
    name: Unlicense
    url: https://opensource.org/licenses/Unlicense
  description: |
    Versioned JSON API for OpenPGP key management.

    **Versioning:** `Accept: application/json; version=1`
    **Auth:** Bearer JWT (Auth0) on `/api/keys/**`, `/api/groups/**`, `/api/invites/**`, and `/api/admin/**`
    **Crypto:** Bouncy Castle (`bcprov-jdk18on`, `bcpg-jdk18on` 1.84). Primary generation supports `openpgpVersion` 4 (default) or 6.
    **Errors:** RFC 7807 `application/problem+json`
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/groups:
    get:
      tags: [Groups]
      summary: List groups
      operationId: listGroups
      x-implementation-status: implemented
      responses:
        '200':
          description: Group list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupResponse'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Groups]
      summary: Create group
      operationId: createGroup
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/groups/{groupId}:
    get:
      tags: [Groups]
      summary: Get group
      operationId: getGroup
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
    patch:
      tags: [Groups]
      summary: Update group
      operationId: updateGroup
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Groups]
      summary: Delete group
      operationId: deleteGroup
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '204':
          description: Group deleted
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/groups/{groupId}/members:
    get:
      tags: [Groups]
      summary: List group members
      operationId: listGroupMembers
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Group member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupMemberResponse'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/groups/{groupId}/members/{memberUserId}:
    delete:
      tags: [Groups]
      summary: Remove group member
      operationId: removeGroupMember
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/MemberUserId'
      responses:
        '204':
          description: Member removed
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/groups/{groupId}/members/me:
    delete:
      tags: [Groups]
      summary: Leave group
      operationId: leaveGroup
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '204':
          description: Current user removed from group
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/groups/{groupId}/invites:
    get:
      tags: [Groups]
      summary: List group invites
      operationId: listGroupInvites
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Pending invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupInviteResponse'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Groups]
      summary: Invite group member
      operationId: inviteGroupMember
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupInviteRequest'
      responses:
        '201':
          description: Invite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupInviteResponse'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/groups/{groupId}/invites/{inviteId}:
    delete:
      tags: [Groups]
      summary: Revoke pending invite
      operationId: revokeGroupInvite
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/InviteId'
      responses:
        '204':
          description: Invite revoked
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/groups/{groupId}/summary:
    get:
      tags: [Groups]
      summary: Get group summary
      operationId: getGroupSummary
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Group summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupSummaryResponse'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/groups/{groupId}/members/audit.csv:
    get:
      tags: [Groups]
      summary: Export group members audit CSV
      operationId: exportGroupMembersAudit
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Group members audit CSV
          content:
            text/csv:
              schema:
                type: string
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: RFC 7807 problem details (typical 400, 401, 404, or 409)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  parameters:
...
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MemberUserId:
      name: memberUserId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    InviteId:
      name: inviteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    ProblemDetail:
      type: object
      description: RFC 7807 Problem Details (Spring `ProblemDetail` JSON)
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: array
          description: Field validation errors (400 responses)
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    GroupMembershipRole:
      type: string
      enum: [owner, member]
...
    GroupResponse:
      type: object
      required: [id, name, ownerUserId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: [string, 'null']
        ownerUserId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateGroupRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 1024
    UpdateGroupRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 1024
    GroupMemberResponse:
      type: object
      required: [groupId, userId, role, joinedAt]
      properties:
        groupId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/GroupMembershipRole'
        invitedByUserId:
          type: [string, 'null']
          format: uuid
        joinedAt:
          type: string
          format: date-time
    CreateGroupInviteRequest:
      type: object
      properties:
        email:
          type: string
          maxLength: 255
        inviteeUserId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/GroupMembershipRole'
        expiresAt:
          type: string
          format: date-time
    GroupInviteResponse:
      type: object
      required: [id, groupId, token, role, invitedByUserId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        token:
          type: string
        email:
          type: [string, 'null']
        inviteeUserId:
          type: [string, 'null']
          format: uuid
        role:
          $ref: '#/components/schemas/GroupMembershipRole'
        invitedByUserId:
          type: string
          format: uuid
        expiresAt:
          type: [string, 'null']
          format: date-time
        acceptedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
    GroupSummaryResponse:
      type: object
      required: [group, memberCount, pendingInviteCount, keyCount]
      properties:
        group:
          $ref: '#/components/schemas/GroupResponse'
        memberCount:
          type: integer
        pendingInviteCount:
          type: integer
        keyCount:
          type: integer
...